ci: add .xmake and jenkins pipeline config
This commit contains internal configuration for build pipelines that do
not exist in the externally available version of NeonBee.

Therefore, an open-source branch has been introduced which exactly
mirrors the version of NeonBee from github.com. The master branch
additionally contains "this" commit, in order to make the build
pipelines on Jenkins and the xmake work properly.